requirements.txt
opencv-python~=4.5.1.48
Pillow~=8.0.0
matplotlib~=3.3.2
networkx~=2.5